feat: obligatorio cambio de credenciales al primer inicio y cifrado con BCrypt
- Implementada la lógica de verificación de primer inicio de sesión en el flujo de autenticación.
- Integrada la librería jBCrypt para el hashing seguro de contraseñas.
- Añadida clase de utilidad SecurityUtils para el manejo de cifrado y validación.
- Actualizado el modelo de Usuario para soportar el estado 'isFirstLogin'.
- Ajustada la navegación en LoginController para redirigir a la vista de configuración inicial.

+++ src/main/java/org/avalon/desktop/product/application/service/InventoryService.java
@@ -36,7 +36,7 @@ public class InventoryService {
     }
 
     public void adjustStock(Long id, int amount) {
-        productRepository.updateStock(id, amount);
+        productRepository.updateStock(id, (double) amount);
     }
 
     public InventoryStats getStats() {
